v3.0-main/requirements.txt
# AI音频转录与会议纪要生成系统依赖包

# 核心框架
flask==2.3.3
flask-cors==4.0.0
# tkinter-tooltip==2.0.0  # 可能不需要，因为我们不使用GUI

# 音频处理
faster-whisper==0.9.0
# pyaudiowpatch==0.2.12.4  # 移除此依赖，因为安装困难
playsound==1.3.0

# AI和语音服务
elevenlabs==0.2.26
requests==2.31.0

# 自动化和图像处理
selenium==4.15.0
opencv-python==4.8.1.78
pyautogui==0.9.54
numpy==1.24.3

# 文件处理
werkzeug==2.3.7
python-dotenv==1.0.0

# 系统工具
psutil==5.9.6
# threading-timer==0.1.0  # 移除此依赖，因为找不到包
# tempfile-utils==0.1.0  # 移除此依赖，因为找不到包

# 开发和调试
pytest==7.4.3
pytest-cov==4.1.0
black==23.9.1
flake8==6.1.0